Successfully retrieved data for Cloudwatch metrics (not modular) with user input
